Model layer created with main business models. View layer created and refactored.
