+++ MyApplication/app/src/main/java/anjelloatoz/com/nodescode/Model/Movie.java
@@ -1,0 +1,12 @@
+package anjelloatoz.com.nodescode.Model;
+
+/**
+ * Created by Anjelloatoz on 8/15/18.
+ */
+
+public class Movie {
+    public int id;
+    public String title;
+    public String overview;
+    public String release_date;
+}
