+++ MyApplication/app/src/main/java/anjelloatoz/com/nodescode/Model/SearchResult.java
@@ -1,0 +1,14 @@
+package anjelloatoz.com.nodescode.Model;
+
+import java.util.ArrayList;
+
+/**
+ * Created by Anjelloatoz on 8/15/18.
+ */
+
+public class SearchResult {
+    public int page;
+    public int total_results;
+    public int total_pages;
+    public ArrayList<Movie> results;
+}
